+++ app/config/uri.js
@@ -1,7 +1,7 @@
 const { mountURL } = require('./config-mapping');
 
-const WORKPLACE_CONTACT_URL = `${mountURL}workplace-contact`;
-const WORKPLACE_CONTACT_CONTEXT_PATH = '/workplace-contact';
+const WORKPLACE_CONTACT_URL = `${mountURL}confirmer`;
+const WORKPLACE_CONTACT_CONTEXT_PATH = '/confirmer';
 const AUTH_URL = `${mountURL}auth`;
 const AUTH_CONTEXT_PATH = '/auth';
 const PUBLIC_URL = `${mountURL}public`;